+++ src/ifs/ext_image_copy.rs
@@ -1,0 +1,4 @@
+pub mod ext_image_copy_capture_cursor_session_v1;
+pub mod ext_image_copy_capture_frame_v1;
+pub mod ext_image_copy_capture_manager_v1;
+pub mod ext_image_copy_capture_session_v1;